// frontend/src/api/model/genre.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */

/**
 * 电影类型实体
 */
export interface Genre {
  /** 类型ID */
  id?: number;
  /** 类型名称 */
  name?: string;
  /** 英文名称 */
  nameEn?: string;
  /** 类型描述 */
  description?: string;
  /** 创建时间 */
  createTime?: string;
}